+++ speaktome/core/beam_tree_node.py
@@ -6,7 +6,7 @@ try:
 except ModuleNotFoundError:  # pragma: no cover - optional
     torch = None
 
-from .tensor_abstraction import (
+from ..tensors import (
     AbstractTensorOperations,
     get_tensor_operations,
 )
