+++ speaktome/core/compressed_beam_tree.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
 
 # Local application/library specific imports
 from .beam_tree_node import BeamTreeNode  # Assuming BeamTreeNode is in beam_tree_node.py
-from .tensor_abstraction import (
+from ..tensors import (
     AbstractTensorOperations,
     get_tensor_operations,
 )
