+++ speaktome/core/lookahead_controller.py
@@ -5,7 +5,7 @@ from typing import List, Tuple, Callable, Any, Set, TYPE_CHECKING
 # Local application/library specific imports
 if TYPE_CHECKING:  # pragma: no cover - type hints only
     from .beam_search_instruction import BeamSearchInstruction
-from .tensor_abstraction import AbstractTensorOperations
+from ..tensors import AbstractTensorOperations
 from .model_abstraction import AbstractModelWrapper
 # --- END HEADER ---
