+++ speaktome/cpu_demo.py
@@ -20,7 +20,7 @@ except ModuleNotFoundError:  # pragma: no cover - optional dependency
     np = None  # type: ignore
 
 from .util.token_vocab import TokenVocabulary
-from .core.tensor_abstraction import get_tensor_operations
+from .tensors import get_tensor_operations
 from .core.model_abstraction import AbstractModelWrapper
 from .core.lookahead_controller import LookaheadController, LookaheadConfig
 # --- END HEADER ---
