+++ speaktome/tensors/torch_backend.py
@@ -1,0 +1,140 @@
+"""PyTorch implementation of :class:`AbstractTensorOperations`."""
+
+from typing import Any, Tuple, List, Optional, Union
+
+try:
+    import torch
+    import torch.nn.functional as F
+except ModuleNotFoundError:  # pragma: no cover - optional dependency
+    torch = None  # type: ignore
+    F = None  # type: ignore
+
+from .abstraction import AbstractTensorOperations
+
+# --- END HEADER ---
+
+class PyTorchTensorOperations(AbstractTensorOperations):
+    def __init__(self, default_device: Union[str, "torch.device"] = "cpu", track_time: bool = False):
+        super().__init__(track_time=track_time)
+        if torch is None:
+            raise RuntimeError("PyTorch is required for this backend")
+        self.default_device = torch.device(default_device)
+
+    def full(self, size, fill_value, dtype, device):
+        return torch.full(size, fill_value, dtype=dtype, device=device or self.default_device)
+
+    def zeros(self, size, dtype, device):
+        return torch.zeros(size, dtype=dtype, device=device or self.default_device)
+
+    def clone(self, tensor):
+        return tensor.clone()
+
+    def to_device(self, tensor, device):
+        return tensor.to(device or self.default_device)
+
+    def get_device(self, tensor):
+        return tensor.device
+
+    def get_dtype(self, tensor):
+        return tensor.dtype
+
+    def item(self, tensor):
+        return tensor.item()
+
+    def max(self, tensor):
+        return torch.max(tensor)
+
+    def long_cast(self, tensor):
+        return tensor.long()
+
+    def not_equal(self, tensor1, tensor2):
+        return tensor1 != tensor2
+
+    def arange(self, start, end=None, step=1, device=None, dtype=None):
+        if end is None:
+            return torch.arange(start, device=device or self.default_device, dtype=dtype)
+        return torch.arange(start, end, step, device=device or self.default_device, dtype=dtype)
+
+    def select_by_indices(self, tensor, indices_dim0, indices_dim1):
+        return tensor[indices_dim0, indices_dim1]
+
+    def log_softmax(self, tensor, dim):
+        return F.log_softmax(tensor, dim=dim)
+
+    def topk(self, tensor, k, dim):
+        return torch.topk(tensor, k=k, dim=dim)
+
+    def stack(self, tensors, dim=0):
+        return torch.stack(tensors, dim=dim)
+
+    def pad(self, tensor, pad, value=0.0):
+        return F.pad(tensor, pad, value=value)
+
+    def cat(self, tensors, dim=0):
+        return torch.cat(tensors, dim=dim)
+
+    def repeat_interleave(self, tensor, repeats, dim=None):
+        return tensor.repeat_interleave(repeats, dim=dim)
+
+    def view_flat(self, tensor):
+        return tensor.view(-1)
+
+    def assign_at_indices(self, tensor_to_modify, indices_dim0, indices_dim1, values_to_assign):
+        tensor_to_modify[indices_dim0, indices_dim1] = values_to_assign
+
+    def increment_at_indices(self, tensor_to_modify, mask):
+        tensor_to_modify[mask] += 1
+
+    def clamp(self, tensor, min_val=None, max_val=None):
+        return torch.clamp(tensor, min=min_val, max=max_val)
+
+    def shape(self, tensor):
+        return tuple(tensor.shape)
+
+    def numel(self, tensor):
+        return tensor.numel()
+
+    def mean(self, tensor, dim=None):
+        return torch.mean(tensor, dim=dim)
+
+    def pow(self, tensor, exponent: float):
+        return torch.pow(tensor, exponent)
+
+    def sqrt(self, tensor):
+        return torch.sqrt(tensor)
+
+    def tensor_from_list(self, data, dtype, device):
+        return torch.tensor(data, dtype=dtype, device=device or self.default_device)
+
+    def boolean_mask_select(self, tensor, mask):
+        return tensor[mask]
+
+    def tolist(self, tensor):
+        return tensor.tolist()
+
+    def less(self, tensor, value):
+        return tensor < value
+
+    def index_select(self, tensor, dim, indices):
+        return torch.index_select(tensor, dim, indices)
+
+    def save(self, tensor, filepath: str) -> None:
+        torch.save(tensor, filepath)
+
+    def load(self, filepath: str, dtype, device):
+        t = torch.load(filepath, map_location=device or self.default_device)
+        if dtype is not None:
+            t = t.to(dtype)
+        return t
+
+    @property
+    def long_dtype(self):
+        return torch.long
+
+    @property
+    def bool_dtype(self):
+        return torch.bool
+
+    @property
+    def float_dtype(self):
+        return torch.float32
